requirements.txt
# Dev environment
pip
autopep8

# App
Flask
google-generativeai
google-api-python-client
firebase-admin
pytz
tqdm
python-telegram-bot
python-telegram-bot==21.4 #when you are using python3